move handleRefreshTasks to App
